requirements.txt
aiogram==3.15.0
Pillow==10.4.0
python-dotenv==1.0.1
